2019-10-21 | First commit

+++ 03_django/03_django_form/articles/views.py
@@ -1,41 +1,43 @@
 from IPython import embed
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from django.views.decorators.http import require_POST
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 
 # Create your views here.
 def index(request):
+    # session 에 visit_num 키로 접근해 값을 가져온다.
+    # 기본적으로 존재하지 않는 키이기 때문에 키가 없다면(방문한적이 없다면) 0 값을 가져오도록 한다.
+    visits_num = request.session.get('visits_num', 0)
+    # 그리고 가져온 값은 session 에 visits_num 에 매번 1씩 증가한 값으로 할당한다. (유저의 다음 방문을 위해)
+    request.session['visits_num'] = visits_num + 1
+    # session data 안에 있는 새로운 정보를 수정했따면 django 는 수정한 사실을 알아채지 못하기 때문에 다음과 같이 설정.
+    request.session.modified = True
+
+
     articles = Article.objects.all()
-    context = {'articles': articles,}
+    context = {'articles': articles, 'visits_num': visits_num,}
     return render(request, 'articles/index.html', context)
 
-
+@login_required
 def create(request):
     if request.method == 'POST':
-        # -----------------필요없음--------------------
-        # title = request.POST.get('title') django form을 이용할 경우 필요없다
-        # content = request.POST.get('content')
-        # article = Article(title=title, content=content)
-        # article.save()
-        # -----------------여기까지 필요 없음-----------------
-
-        # Form Instance를 생성하고 요청에 의한 데이터를 인자로 받는다.(binding)
-        # 이 처리과정은 binding이라고 불리며 유효성 체크를 할 수 있게 해준다.
+        # form 인스턴스를 생성하고 요청에 의한 데이터를 인자로 받는다. (binding)
+        # 이 처리과정은 binding 이라고 불리며 유효성 체크를 할 수 있도록 해준다.
         form = ArticleForm(request.POST)
-        if form.is_valid(): # 유효성 검사
+        # form 이 유효한지 체크한다.
+        if form.is_valid():
             article = form.save()
             return redirect(article)
-    # 상황에 따라 context에 다른 form을 넘겨주어야 한다.
-    # GET : 기본 form
-    # POST : 검증에 실패 후의 form(is_valid = False일때)
-    # POST로 들어왔을때 검증이 실패할 시, 경로를 잡아주기 위해 context와 return은 if와 else와 같은 선상에 작성.
     else:
         form = ArticleForm()
+    # 상황에 따라 context 에 넘어가는 2가지 form
+    # 1. GET : 기본 form
+    # 2. POST : 검증에 실패 후의 form(is_valid == False)
     context = {'form': form,}
     return render(request, 'articles/form.html', context)
-
 
 
 def detail(request, article_pk):
@@ -48,11 +50,13 @@ def detail(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
 
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'POST':
@@ -68,17 +72,20 @@ def update(request, article_pk):
 
 @require_POST
 def comments_create(request, article_pk):
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        # 객체를 Create 하지만, db 에 레코드는 작성하지 않는다.
-        comment =  comment_form.save(commit=False)
-        comment.article_id = article_pk
-        comment.save()
+    if request.user.is_authenticated:
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            # 객체를 Create 하지만, db 에 레코드는 작성하지 않는다.
+            comment = comment_form.save(commit=False)
+            comment.article_id = article_pk
+            comment.save()
     return redirect('articles:detail', article_pk)
 
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
+        return redirect('articles:detail', article_pk)
+    return HttpResponse('You are Unauthorized', status=401)
